refactor query
