+++ libs/data-access/middleware-service-worker/src/lib/convertToUTC.ts
@@ -1,0 +1,6 @@
+export function convertToUTC(date?: string) {
+  if (date && !date.endsWith('Z')) {
+    return date + 'Z';
+  }
+  return date;
+}
